requirements.txt
streamlit
google-generativeai>=0.5.0
PyPDF2
gspread
oauth2client